Reorganize entities query code to simplify/clarify
