+++ src/pages/EntitiesSearchPage/EntitiesSearchPage.css
