+++ src/pages/EntitiesSearchPage/entitiesQueryBuilder.js
